requirements-dev.txt
# If there are tools that a developer needs to work on your project.
# Unit testing tools, mocking frameworks, debuggers, proxies, ...
# These will be used for tox and other testing environments.
-r requirements.txt
flake8==2.4.0
Flask-WebTest==0.0.7
ipdb
mock==1.3.0
NoseGAE==0.5.8
